Full screen - corrections
