+++ packages/client/src/components/FullscreenProvider/index.css
@@ -1,0 +1,4 @@
+.FullScreen {
+  width: 100%;
+  height: 100%;
+}
